minor refacto
