+++ app/helpers/users_filters_helper.rb
@@ -1,0 +1,40 @@
+module UsersFiltersHelper
+  def filter_list
+    [
+      :search_query,
+      :tag_ids,
+      :status,
+      :orientation_type,
+      :action_required,
+      :referent_id,
+      :creation_date_after,
+      :creation_date_before,
+      :convocation_date_before,
+      :convocation_date_after,
+      :first_invitation_date_before,
+      :first_invitation_date_after,
+      :last_invitation_date_before,
+      :last_invitation_date_after
+    ]
+  end
+
+  def active_filter_list
+    filter_list.select { |filter| params[filter].present? }
+  end
+
+  def invitation_or_convocation_active_filters_count
+    active_filters_count = 0
+
+    active_filters_count += 1 if params[:convocation_date_after].present? || params[:convocation_date_before].present?
+
+    if params[:first_invitation_date_after].present? || params[:first_invitation_date_before].present?
+      active_filters_count += 1
+    end
+
+    if params[:last_invitation_date_after].present? || params[:last_invitation_date_before].present?
+      active_filters_count += 1
+    end
+
+    active_filters_count
+  end
+end
