+++ app/views/users/_active_filters_recap.html.erb
@@ -6,33 +6,23 @@
     <% end %>
   </div>
   <div class="active-filters-container d-flex flex-wrap">
-    <% (active_filter_list - [:search_query, :creation_date_after, :creation_date_before, :convocation_date_before, :convocation_date_after, :first_invitation_date_before, :first_invitation_date_after, :last_invitation_date_before, :last_invitation_date_after]).each do |filter| %>
-      <% if filter == :tag_ids %>
-        <% @filtered_tags.each do |tag| %>
-          <span class="d-flex me-2 active-filter-badge">
-            <%= tag.value %>
-            <%= link_to structure_users_path(url_params.merge(tag_ids: params[:tag_ids] - [tag.id.to_s])) do %>
-              <i class="ri-close-line"></i>
-            <% end %>
-          </span>
+    <% ([:status, :orientation_type, :action_required, :referent_id] & active_filter_list).each do |filter| %>
+      <span class="d-flex me-2 active-filter-badge">
+        <% if filter == :referent_id %>
+          Suivi par <%= @referent %>
+        <% elsif filter == :orientation_type %>
+          Orientation : <%= params[filter] %>
+        <% elsif filter == :action_required %>
+          Intervention nécessaire
+        <% elsif filter == :status %>
+          Statut : <%= I18n.t("activerecord.attributes.follow_up.statuses.#{params[filter]}") %>
         <% end %>
-      <% else %>
-        <span class="d-flex me-2 active-filter-badge">
-          <% if filter == :referent_id %>
-            Suivi par <%= @referent %>
-          <% elsif filter == :orientation_type %>
-            Orientation : <%= params[filter] %>
-          <% elsif filter == :action_required %>
-            Intervention nécessaire
-          <% elsif filter == :status %>
-            Statut : <%= I18n.t("activerecord.attributes.follow_up.statuses.#{params[filter]}") %>
-          <% end %>
-          <%= link_to structure_users_path(url_params.except(filter)) do %>
-            <i class="ri-close-line"></i>
-          <% end %>
-        </span>
-      <% end %>
+        <%= link_to structure_users_path(url_params.except(filter)) do %>
+          <i class="ri-close-line"></i>
+        <% end %>
+      </span>
     <% end %>
+    <%= render "users/active_filters_recap/tag_ids" %>
     <%= render "users/active_filters_recap/creation_date" %>
     <%= render "users/active_filters_recap/convocation_date" %>
     <%= render "users/active_filters_recap/first_invitation_date" %>
