+++ app/views/users/active_filters_recap/_tag_ids.html.erb
@@ -1,0 +1,10 @@
+<% if params[:tag_ids] %>
+  <% @filtered_tags.each do |tag| %>
+    <span class="d-flex me-2 active-filter-badge">
+      <%= tag.value %>
+      <%= link_to structure_users_path(url_params.merge(tag_ids: params[:tag_ids] - [tag.id.to_s])) do %>
+        <i class="ri-close-line"></i>
+      <% end %>
+    </span>
+  <% end %>
+<% end %>
